+++ darkness-forge-1.12.x/CHANGELOG.md
@@ -1,2 +1,2 @@
-- Fixed an issue when loading metadata from (wrong) mcmod.info file would cause a crash at startup.
-  See https://github.com/yamporg/mods/issues/1
+- Fixed loading resource pack with config localization.   
+  See https://github.com/yamporg/mods/issues/6